Downloaded a filtered out download fix
Found that if a download has been found and denied due to a filter other sites may have what you want.
Code now looks for the title, sees if it was denied, and if the newly found can download it will download.
This should allow for better integration of multiple websites offers the same things in title but at different filter settings.