Fixed multiple issues and added ui color
Fixed download button being triggered by clicking on the edit filter button by adding a tiny delay
Fixed major error in which changing options in the middle of a parsing causes the download to do unintended things.

Added a red tint that shows that multiple options can not be changed when a parsing is happening.
(BUG: Red tint may flash for a few seconds after a complete parse)